pyproject.toml
[project]
name = "agents-config"
version = "0.1.0"
description = "Robust Pydantic models for AI agent configuration with environment variable substitution"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "kdcllc", email = "info@kingdavidconsulting.com" },
]
maintainers = [
    { name = "kdcllc", email = "info@kingdavidconsulting.com" },
]
license = "MIT"
keywords = ["ai", "agents", "configuration", "pydantic", "yaml"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = ["pydantic>=2.0.0", "PyYAML>=6.0.0", "typing-extensions>=4.0.0"]

[project.urls]
Homepage = "https://github.com/kdcllc/agents_config"
Documentation = "https://github.com/kdcllc/agents_config#readme"
Repository = "https://github.com/kdcllc/agents_config.git"
"Bug Tracker" = "https://github.com/kdcllc/agents_config/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app"]

[tool.black]
line-length = 150
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 150

[tool.flake8]
max-line-length = 150
ignore = ["E203", "W503"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[dependency-groups]
dev = ["build>=1.3.0", "twine>=6.1.0"]
